Create database tables and seeded files

+++ upkeeper-api/server/db/schema/04_create_cleaner.sql
@@ -3,7 +3,7 @@ DROP TABLE IF EXISTS cleaners CASCADE;
 -- CREATE CLEANER
 CREATE TABLE cleaners (
   id SERIAL PRIMARY KEY,
-  payrate INTEGER(300) NOT NULL,
+  payrate INTEGER NOT NULL,
   user_id integer REFERENCES users(id) ON DELETE CASCADE NOT NULL,
   location_id integer REFERENCES locations(id) ON DELETE CASCADE NOT NULL
 );
